requirements.txt
prometheus-client==0.16.0
psycopg2-binary==2.9.4
requests==2.28.2